+++ packages/fcl/src/current-user/url-from-service.js
@@ -1,3 +1,5 @@
+import {URL} from "../utils/url"
+
 export function urlFromService(service, includeParams = true) {
   const url = new URL(service.endpoint)
   if (includeParams) {