+++ packages/fcl/src/discovery/services.js
@@ -1,9 +1,10 @@
 import {config} from "@onflow/config"
 import {invariant} from "@onflow/util-invariant"
-import {serviceRegistry} from "../current-user/exec-service/plugins"
+import {getServiceRegistry} from "../current-user/exec-service/plugins"
 import {getChainId} from "../utils"
 import {VERSION} from "../VERSION"
 import {makeDiscoveryServices} from "./utils"
+import {URL} from "../utils/url"
 
 export async function getServices({types}) {
   const endpoint = await config.get("discovery.authn.endpoint")
@@ -25,7 +26,7 @@ export async function getServices({types}) {
       fclVersion: VERSION,
       include,
       clientServices: await makeDiscoveryServices(),
-      supportedStrategies: serviceRegistry.getStrategies(),
+      supportedStrategies: getServiceRegistry().getStrategies(),
       userAgent: window?.navigator?.userAgent,
       network: await getChainId(),
     }),